% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf0.R
\name{nmf0}
\alias{nmf0}
\title{Fit a Fast Scalable Non-negative Matrix Factorization using L-0 penalization}
\usage{
nmf0(Xmat, K = 1L, knownF, lambda1 = 0.1, lambda2 = 0.1,
  lambda3 = 0.1, tol = 1e-04, maxiter = 1000, verb = TRUE,
  init_method = "random", semi = FALSE, hard_keep = NULL)
}
\arguments{
\item{K}{the number of clusters to fit. Must be an integer.}

\item{knownF}{The number of known factors to be used. Default is missing in which case the unsupervised NMF0
method is used.}

\item{lambda1}{The tuning parameter for the L-0 penalty on the loading matrix L}

\item{lambda2}{The tuning parameter for the L-2 penalty on the loading matrix L}

\item{lambda3}{The tuning parameter for the L-2 penalty on the factor matrix F}

\item{tol}{The relative tolerance which when met calls for stoppage in the optimization run}

\item{maxiter}{The maximum number of iterations for which to run the iterative updates in nmf0}

\item{verb}{If TRUE, prints the progress of the model fit}

\item{init_method}{The method for initializing the NMF method. Can be one of two values - random and svd.
when init_method=random, the initial updates to L and F are decided randomly}

\item{semi}{If FALSE, then enforces that the factor matrix is non-negative, else does not make that assumption,
Default is FALSE.}

\item{hard_keep}{The maximum number of clusters that are representative of each sample.}

\item{X}{The input data matrix with rows being the features and columns the samples.}
}
\value{
Outputs the best NMF0 fitted model for cluster K that includes the estimated L and F matrices, and the
        model likelihood.
}
\description{
Fits a grade of membership model using non-negative matrix factorization using L-0 penalization
             that can handle large scale input data and is fast and scalable. Also allows for unsupervised
             and semi-supervised set ups for factors/clusters.
             X=LF
}
\examples{

data("ex.counts")
out <- nmf0(t(ex.counts), K=4, tol=0.001, maxiter=50)
}
\references{
Hussein Hazimeh and Rahul Mazumder.2018. Fast Best Subset Selection:
             Coordinate Descent and Local Combinatorial Optimization Algorithms.
             arXiv preprint arXiv:1803.01454.
}
